+++ packages/backend/src/services/ServiceRepository.ts
@@ -16,6 +16,7 @@ import { GithubAppService } from './GithubAppService/GithubAppService';
 import { GitIntegrationService } from './GitIntegrationService/GitIntegrationService';
 import { GroupsService } from './GroupService';
 import { HealthService } from './HealthService/HealthService';
+import { MetricsExplorerService } from './MetricsExplorerService/MetricsExplorerService';
 import { NotificationsService } from './NotificationsService/NotificationsService';
 import { OrganizationService } from './OrganizationService/OrganizationService';
 import { PersonalAccessTokenService } from './PersonalAccessTokenService';
@@ -70,6 +71,7 @@ interface ServiceManifest {
     userService: UserService;
     validationService: ValidationService;
     catalogService: CatalogService;
+    metricsExplorerService: MetricsExplorerService;
     promoteService: PromoteService;
     savedSqlService: SavedSqlService;
     contentService: ContentService;
@@ -652,6 +654,20 @@ export class ServiceRepository
                     savedChartModel: this.models.getSavedChartModel(),
                     spaceModel: this.models.getSpaceModel(),
                     tagsModel: this.models.getTagsModel(),
+                }),
+        );
+    }
+
+    public getMetricsExplorerService(): MetricsExplorerService {
+        return this.getService(
+            'metricsExplorerService',
+            () =>
+                new MetricsExplorerService({
+                    lightdashConfig: this.context.lightdashConfig,
+                    catalogModel: this.models.getCatalogModel(),
+                    projectService: this.getProjectService(),
+                    catalogService: this.getCatalogService(),
+                    projectModel: this.models.getProjectModel(),
                 }),
         );
     }